# task4/docs/swagger.yaml
basePath: /api/v1
definitions:
  user.CreateUserRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    required:
    - first_name
    - last_name
    type: object
  user.UpdateUserRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    type: object
  user.User:
    properties:
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the example of CRUD API using go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User CRUD API
  version: "1.0"
paths:
  /user:
    post:
      consumes:
      - application/json
      description: Create a new user with specified First and Last names
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new user
      tags:
      - user
  /user/{id}:
    delete:
      description: Delete User By Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete User By Id
      tags:
      - user
    get:
      description: Get User Object By Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Find User By Id
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update a Last name or a First name of the specific user
      parameters:
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update a user
      tags:
      - user
swagger: "2.0"